refactor: rename getPaths -> validFormPaths
